refactor(scout-for-lol): move explore lib modules into their own sub-domain
Why:
main sits exactly at the directory-file-count ceiling (69) for
app/src/lib, so this branch's new explore-mentions.ts/.test.ts tips it to
70 and fails //#check-directory-file-counts. The ceiling is a ratchet with
no allowlist — the fix is splitting the directory, not raising it.

What:
git mv every explore-*.ts module (and its colocated test) from
app/src/lib into app/src/lib/explore, matching the components/explore
split already done elsewhere, and update the ~14 importers.

Verification:
bun run scripts/check-directory-file-counts.ts reports no errors (lib
drops to 58 source files).
bunx turbo run build typecheck test lint --filter="./packages/scout-for-lol/packages/*"
passes in full (63/63 tasks).
