+++ packages/scout-for-lol/packages/app/src/components/explore/explore-creation-intent-card.tsx
@@ -15,7 +15,7 @@ import {
   creationConfirmedEvent,
   createdEntityLink,
   type CreationIntentCardData,
-} from "#src/lib/explore-intent-cards.ts";
+} from "#src/lib/explore/explore-intent-cards.ts";
 import {
   classifyCreationIntentConfirmation,
   type CreationConfirmationOutcome,
