+++ packages/scout-for-lol/packages/app/src/hooks/use-explore-run-markers.ts
@@ -7,7 +7,7 @@ import {
   loadExploreRunMarkers,
   saveExploreRunMarkers,
   type ExploreRunMarker,
-} from "#src/lib/explore-run-markers.ts";
+} from "#src/lib/explore/explore-run-markers.ts";
 
 function initialMarkers(): ExploreRunMarker[] {
   return loadExploreRunMarkers(globalThis.localStorage);
