+++ packages/scout-for-lol/packages/app/src/lib/explore/explore-export.test.ts
@@ -3,7 +3,7 @@ import { ExploreMessageSchema, type ExploreMessage } from "@scout-for-lol/data";
 import {
   conversationToMarkdown,
   exportFilename,
-} from "#src/lib/explore-export.ts";
+} from "#src/lib/explore/explore-export.ts";
 
 const QUESTION_ID = "33333333-3333-4333-8333-333333333333";
 const ANSWER_ID = "44444444-4444-4444-8444-444444444444";
