+++ packages/scout-for-lol/packages/app/src/lib/explore/explore-intent-cards.ts
