+++ packages/scout-for-lol/packages/app/src/lib/explore/explore-navigation.ts
