+++ packages/scout-for-lol/packages/app/src/lib/explore/explore-turn-state.ts
